Extract a real base class out of `Saver`.
